+++ app/models/automated_checks/claim_verifier.rb
@@ -2,24 +2,24 @@ module AutomatedChecks
   class ClaimVerifier
     def initialize(
       claim:,
-      dqt_teacher_status:,
+      dqt_teacher_statuses:,
       admin_user: nil,
       verifiers: [
         ClaimVerifiers::Identity.new(
           admin_user: admin_user,
           claim: claim,
-          dqt_teacher_status: dqt_teacher_status
+          dqt_teacher_statuses: dqt_teacher_statuses
         ),
         ClaimVerifiers::Qualifications.new(
           admin_user: admin_user,
           claim: claim,
-          dqt_teacher_status: dqt_teacher_status
+          dqt_teacher_statuses: dqt_teacher_statuses
         )
       ]
     )
       self.admin_user = admin_user
       self.claim = claim
-      self.qualified_teaching_status = dqt_teacher_status
+      self.qualified_teaching_statuses = dqt_teacher_statuses
       self.verifiers = verifiers
     end
 
@@ -31,6 +31,6 @@ module AutomatedChecks
 
     private
 
-    attr_accessor :admin_user, :claim, :qualified_teaching_status, :verifiers
+    attr_accessor :admin_user, :claim, :qualified_teaching_statuses, :verifiers
   end
 end