+++ app/models/claim.rb
@@ -90,7 +90,9 @@ class Claim < ApplicationRecord
     personal_data_removed_at: false,
     email_verified: false,
     one_time_password: true,
-    sent_one_time_password_at: false
+    sent_one_time_password_at: false,
+    mobile_verified: false,
+    one_time_password_category: false
   }.freeze
   DECISION_DEADLINE = 12.weeks
   DECISION_DEADLINE_WARNING_POINT = 2.weeks
@@ -192,12 +194,15 @@ class Claim < ApplicationRecord
 
   validates :email_address, on: [:"email-address", :submit], presence: {message: "Enter an email address"}
   validates :email_address, format: {with: URI::MailTo::EMAIL_REGEXP, message: "Enter an email in the format name@example.com"},
-                            length: {maximum: 256, message: "Email address must be 256 characters or less"},
-                            allow_blank: true
+                            length: {maximum: 256, message: "Email address must be 256 characters or less"}, if: -> { email_address.present? }
 
   validates :provide_mobile_number, on: [:"provide-mobile-number", :submit], inclusion: {in: [true, false], message: "Choose yes if you want to provide a mobile number"}, if: :has_ecp_policy?
-  validates :mobile_number, on: [:"mobile-number"], presence: {message: "Enter a mobile number in the correct format, for example 07123456789"}, if: :has_ecp_policy?
-  validates :mobile_number, on: [:"mobile-number"], format: {with: /\A\+?(?:\d\s?){11}\z/, message: "A mobile number must be 11 digits"}, if: :has_ecp_policy?
+  validates :mobile_number, on: [:"mobile-number", :submit], presence: {message: "Enter a mobile number in the correct format, for example 07123456789"}, if: -> { provide_mobile_number == true && has_ecp_policy? }
+  validates :mobile_number,
+    format: {
+      with: /\A\+?(?:\d\s?){11}\z/,
+      message: "A mobile number must be 11 digits"
+    }, if: -> { provide_mobile_number == true && mobile_number.present? && has_ecp_policy? }
 
   validates :bank_or_building_society, on: [:"bank-or-building-society", :submit], presence: {message: "Select personal bank account or building society"}
   validates :banking_name, on: [:"personal-bank-account", :"building-society-account", :submit], presence: {message: "Enter a name on the account"}