+++ lib/fwy/objects/teacher.rb
@@ -5,27 +5,29 @@ module Fwy
     end
 
     def first_name
-      name&.split&.first
+      string_reader(name) do |string|
+        string.split.first
+      end
     end
 
     def surname
-      name&.split&.last
+      string_reader(name) do |string|
+        string.split.last
+      end
     end
 
-    def qts_date
-      qualified_teacher_status&.qts_date&.to_date
+    def qts_award_date
+      date_reader(qualified_teacher_status&.qts_date)
     end
 
     def itt_subject_codes
-      [
-        itt&.subject1,
-        itt&.subject2,
-        itt&.subject3
-      ].compact
+      (1..3).filter_map do |n|
+        string_reader(itt&.send("subject#{n}"))
+      end
     end
 
-    def itt_date
-      itt.presence&.programme_start_date&.to_date
+    def itt_start_date
+      date_reader(itt.presence&.programme_start_date)
     end
 
     def qualification_name
@@ -46,6 +48,10 @@ module Fwy
     
     def itt
       initial_teacher_training
+    end
+
+    def active_alert?
+      boolean_reader(active_alert)
     end
   end
 end