+++ spec/features/maths_and_physics_claim_spec.rb
@@ -155,26 +155,13 @@ RSpec.feature "Maths & Physics claims" do
       # - Provide mobile number
       expect(page).to have_text(I18n.t("questions.provide_mobile_number"))
 
-      choose "Yes"
+      choose "No"
       click_on "Continue"
 
-      expect(claim.reload.provide_mobile_number).to eql true
+      expect(claim.reload.provide_mobile_number).to eql false
 
       # - Mobile number
-      expect(page).to have_text(I18n.t("questions.mobile_number"))
-
-      fill_in "claim_mobile_number", with: "07123456789"
-      click_on "Continue"
-
-      expect(claim.reload.mobile_number).to eql("07123456789")
-
-      # - Mobile number one-time password
-      # expect(page).to have_text("Password verification")
-      # expect(page).to have_text("Enter the 6-digit password")
-      # expect(page).not_to have_text("We recommend you copy and paste the password from the email.")
-
-      # fill_in "claim_one_time_password", with: otp_sent_to_mobile
-      # click_on "Confirm"
+      expect(page).not_to have_text(I18n.t("questions.mobile_number"))
 
       expect(page).to have_text(I18n.t("questions.bank_or_building_society"))
 
@@ -200,7 +187,13 @@ RSpec.feature "Maths & Physics claims" do
       expect(page).to have_text("Check your answers before sending your application")
 
       stub_geckoboard_dataset_update
-      stub_qualified_teaching_status_show(claim: claim)
+
+      stub_qualified_teaching_statuses_show(
+        query: {
+          trn: claim.teacher_reference_number,
+          ni: claim.national_insurance_number
+        }
+      )
 
       freeze_time do
         perform_enqueued_jobs do