+++ spec/models/maths_and_physics/dqt_record_spec.rb
@@ -18,15 +18,15 @@ RSpec.describe MathsAndPhysics::DqtRecord do
         let(:attributes) { {itt_subject_codes: [jac_code], degree_codes: []} }
 
         it "returns true if the given QTS award date is after the first eligible academic year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("19/3/2017")})).eligible?).to eql true
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("19/3/2017")}))).eligible?).to eql true
         end
 
         it "returns true if the given QTS award date is in the first eligible academic year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("1/10/2015")})).eligible?).to eql true
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("1/10/2015")}))).eligible?).to eql true
         end
 
         it "returns false if the given date is not an eligible year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("8/3/2000")})).eligible?).to eql false
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("8/3/2000")}))).eligible?).to eql false
         end
       end
 
@@ -34,20 +34,20 @@ RSpec.describe MathsAndPhysics::DqtRecord do
         let(:attributes) { {degree_codes: [jac_code], itt_subject_codes: []} }
 
         it "returns true if the given QTS award date is after the first eligible academic year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("19/3/2017")})).eligible?).to eql true
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("19/3/2017")}))).eligible?).to eql true
         end
 
         it "returns true if the given QTS award date is in the first eligible academic year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("1/10/2015")})).eligible?).to eql true
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("1/10/2015")}))).eligible?).to eql true
         end
 
         it "returns false if the given date is not an eligible year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("8/3/2000")})).eligible?).to eql false
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("8/3/2000")}))).eligible?).to eql false
         end
       end
 
-      it "returns false when given a record with a blank qts_date" do
-        expect(MathsAndPhysics::DqtRecord.new({qts_date: "", itt_subject_codes: [jac_code], degree_codes: []}).eligible?).to eql false
+      it "returns false when given a record with a blank qts_award_date" do
+        expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new({qts_award_date: "", itt_subject_codes: [jac_code], degree_codes: []})).eligible?).to eql false
       end
     end
 
@@ -56,15 +56,15 @@ RSpec.describe MathsAndPhysics::DqtRecord do
         let(:attributes) { {itt_subject_codes: [hecos_code], degree_codes: []} }
 
         it "returns true if the given QTS award date is after the first eligible academic year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("1/10/2018")})).eligible?).to eql true
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("1/10/2018")}))).eligible?).to eql true
         end
 
         it "returns true if the given QTS award date is in the first eligible academic year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("1/10/2015")})).eligible?).to eql true
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("1/10/2015")}))).eligible?).to eql true
         end
 
         it "returns false if the given date is not an eligible year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("8/3/2000")})).eligible?).to eql false
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("8/3/2000")}))).eligible?).to eql false
         end
       end
 
@@ -72,20 +72,20 @@ RSpec.describe MathsAndPhysics::DqtRecord do
         let(:attributes) { {degree_codes: [hecos_code], itt_subject_codes: []} }
 
         it "returns true if the given QTS award date is after the first eligible academic year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("19/3/2017")})).eligible?).to eql true
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("19/3/2017")}))).eligible?).to eql true
         end
 
         it "returns true if the given QTS award date is in the first eligible academic year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("1/10/2015")})).eligible?).to eql true
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("1/10/2015")}))).eligible?).to eql true
         end
 
         it "returns false if the given date is not an eligible year" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("8/3/2000")})).eligible?).to eql false
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("8/3/2000")}))).eligible?).to eql false
         end
       end
 
-      it "returns false when given a record with a blank qts_date" do
-        expect(MathsAndPhysics::DqtRecord.new({qts_date: "", itt_subject_codes: [hecos_code], degree_codes: []}).eligible?).to eql false
+      it "returns false when given a record with a blank qts_award_date" do
+        expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new({qts_award_date: "", itt_subject_codes: [hecos_code], degree_codes: []})).eligible?).to eql false
       end
     end
 
@@ -94,7 +94,7 @@ RSpec.describe MathsAndPhysics::DqtRecord do
         let(:attributes) { {itt_subject_codes: [jac_code], degree_codes: [jac_code]} }
 
         it "always returns false" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("19/3/2017")})).eligible?).to eql false
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("19/3/2017")}))).eligible?).to eql false
         end
       end
     end
@@ -103,7 +103,7 @@ RSpec.describe MathsAndPhysics::DqtRecord do
       context "when given ITT subject or degree (#{hecos_code}) - HECOS code isn't eligible" do
         let(:attributes) { {itt_subject_codes: [hecos_code], degree_codes: [hecos_code]} }
         it "always returns false" do
-          expect(MathsAndPhysics::DqtRecord.new(attributes.merge({qts_date: Date.parse("20/07/2017")})).eligible?).to eql false
+          expect(MathsAndPhysics::DqtRecord.new(OpenStruct.new(attributes.merge({qts_award_date: Date.parse("20/07/2017")}))).eligible?).to eql false
         end
       end
     end